#9 Dataの定義ファイルをentitiesに移し、データ構造を再定義
